vbh/build-harness-extensions/modules/clusterpool/requirements.txt
gitpython
paramiko
travispy
python-dateutil
